:art: Refactor functions and files
