+++ services/blockchain-indexer/shared/dataService/dynamicReward/annualInflation.js
